Fix potential memory leaks in aio_free()
aio_free() used to directly free the events, bypassing the add/delete
queue.
If the delete handler of an event tries to add/delete another event, it
will remain in the queue and never be processed.

We fixed the issue by using aio_event_del() and processing the queue,
this however does not prevent leaks from a delete handler creating a new
AIO event.